Module openidconnect::core

Baseline OpenID Connect implementation and types.

Structs

CoreClaimName

OpenID Connect Core claim name.

CoreGenderClaim

OpenID Connect Core gender claim.

CoreHmacKey

HMAC secret key.

CoreJsonWebKey

Public or symmetric key expressed as a JSON Web Key.

CoreRsaPrivateSigningKey

RSA private key.

Enums

CoreApplicationType

OpenID Connect Core client application type.

CoreAuthDisplay

How the Authorization Server displays the authentication and consent user interface pages to the End-User.

CoreAuthErrorResponseType

OpenID Connect Core authentication error response types.

CoreAuthPrompt

Whether the Authorization Server should prompt the End-User for re-authentication and consent.

CoreClaimType

Representation of a Claim Value.

CoreClientAuthMethod

OpenID Connect Core client authentication method.

CoreErrorResponseType

Basic access token error types.

CoreGrantType

OpenID Connect Core grant type.

CoreJsonWebKeyType

Type of JSON Web Key.

CoreJsonWebKeyUse

Usage restriction for a JSON Web key.

CoreJweContentEncryptionAlgorithm

OpenID Connect Core JWE encryption algorithms.

CoreJweKeyManagementAlgorithm

OpenID Connect Core JWE key management algorithms.

CoreJwsSigningAlgorithm

OpenID Connect Core JWS signing algorithms.

CoreRegisterErrorResponseType

OpenID Connect Core registration error response type.

CoreResponseMode

OpenID Connect Core response mode.

CoreResponseType

OpenID Connect Core response type.

CoreRevocableToken

A token representation usable with authorization servers that support RFC 7009 token revocation.

CoreSubjectIdentifierType

OpenID Connect Core Subject Identifier type.

CoreTokenType

Basic OAuth2 authorization token types.

Type Definitions

CoreAuthenticationFlow

OpenID Connect Core authentication flows.

CoreClient

OpenID Connect Core client.

CoreClientMetadata

OpenID Connect Core client metadata.

CoreClientRegistrationRequest

OpenID Connect Core client registration request.

CoreClientRegistrationResponse

OpenID Connect Core client registration response.

CoreIdToken

OpenID Connect Core ID token.

CoreIdTokenClaims

OpenID Connect Core ID token claims.

CoreIdTokenFields

OpenID Connect Core ID token fields.

CoreIdTokenVerifier

OpenID Connect Core ID token verifier.

CoreJsonWebKeySet

OpenID Connect Core JSON Web Key Set.

CoreProviderMetadata

OpenID Connect Core provider metadata.

CoreRequestTokenError

Token error specialization for basic OAuth2 implementation.

CoreRevocationErrorResponse

Revocation error response specialization for basic OAuth2 implementation.

CoreTokenIntrospectionResponse

OpenID Connect Core token introspection response.

CoreTokenResponse

OpenID Connect Core token response.

CoreUserInfoClaims

OpenID Connect Core user info claims.

CoreUserInfoJsonWebToken

OpenID Connect Core user info JSON Web Token.

CoreUserInfoVerifier

OpenID Connect Core user info verifier.