Struct openidconnect::UserInfoClaims

pub struct UserInfoClaims<AC: AdditionalClaims, GC: GenderClaim>(_);

User info claims.

Implementations

impl<AC, GC> UserInfoClaims<AC, GC> where
AC: AdditionalClaims,
GC: GenderClaim,

pub fn new(standard_claims: StandardClaims<GC>, additional_claims: AC) -> Self

Initializes user info claims.

pub fn from_json<RE>(
user_info_json: &[u8],
expected_subject: Option<&SubjectIdentifier>
) -> Result<Self, UserInfoError<RE>> where
RE: Error + 'static,

Initializes user info claims from the provided raw JSON response.

If an expected_subject is provided, this function verifies that the user info claims contain the expected subject and returns an error otherwise.

pub fn issuer(&self) -> Option<&IssuerUrl>

Returns the issuer claim.

pub fn set_issuer(self, issuer: Option<IssuerUrl>) -> Self

Sets the issuer claim.

pub fn audiences(&self) -> Option<&Vec<Audience>>

Returns the aud claim.

pub fn set_audiences(self, audiences: Option<Vec<Audience>>) -> Self

Sets the aud claim.

pub fn subject(&self) -> &SubjectIdentifier

Returns the sub claim.

pub fn set_subject(&mut self, subject: SubjectIdentifier)

Sets the sub claim.

pub fn name(&self) -> Option<&LocalizedClaim<EndUserName>>

Returns the name claim.

pub fn set_name(self, name: Option<LocalizedClaim<EndUserName>>) -> Self

Sets the name claim.

pub fn given_name(&self) -> Option<&LocalizedClaim<EndUserGivenName>>

Returns the given_name claim.

pub fn set_given_name(
self,
given_name: Option<LocalizedClaim<EndUserGivenName>>
) -> Self

Sets the given_name claim.

pub fn family_name(&self) -> Option<&LocalizedClaim<EndUserFamilyName>>

Returns the family_name claim.

pub fn set_family_name(
self,
family_name: Option<LocalizedClaim<EndUserFamilyName>>
) -> Self

Sets the family_name claim.

pub fn middle_name(&self) -> Option<&LocalizedClaim<EndUserMiddleName>>

Returns the middle_name claim.

pub fn set_middle_name(
self,
middle_name: Option<LocalizedClaim<EndUserMiddleName>>
) -> Self

Sets the middle_name claim.

pub fn nickname(&self) -> Option<&LocalizedClaim<EndUserNickname>>

Returns the nickname claim.

pub fn set_nickname(
self,
nickname: Option<LocalizedClaim<EndUserNickname>>
) -> Self

Sets the nickname claim.

pub fn preferred_username(&self) -> Option<&EndUserUsername>

Returns the preferred_username claim.

pub fn set_preferred_username(
self,
preferred_username: Option<EndUserUsername>
) -> Self

Sets the preferred_username claim.

pub fn profile(&self) -> Option<&LocalizedClaim<EndUserProfileUrl>>

Returns the profile claim.

pub fn set_profile(
self,
profile: Option<LocalizedClaim<EndUserProfileUrl>>
) -> Self

Sets the profile claim.

pub fn picture(&self) -> Option<&LocalizedClaim<EndUserPictureUrl>>

Returns the picture claim.

pub fn set_picture(
self,
picture: Option<LocalizedClaim<EndUserPictureUrl>>
) -> Self

Sets the picture claim.

pub fn website(&self) -> Option<&LocalizedClaim<EndUserWebsiteUrl>>

Returns the website claim.

pub fn set_website(
self,
website: Option<LocalizedClaim<EndUserWebsiteUrl>>
) -> Self

Sets the website claim.

pub fn email(&self) -> Option<&EndUserEmail>

Returns the email claim.

pub fn set_email(self, email: Option<EndUserEmail>) -> Self

Sets the email claim.

pub fn email_verified(&self) -> Option<bool>

Returns the email_verified claim.

pub fn set_email_verified(self, email_verified: Option<bool>) -> Self

Sets the email_verified claim.

pub fn gender(&self) -> Option<&GC>

Returns the gender claim.

pub fn set_gender(self, gender: Option<GC>) -> Self

Sets the gender claim.

pub fn birthday(&self) -> Option<&EndUserBirthday>

Returns the birthday claim.

pub fn set_birthday(self, birthday: Option<EndUserBirthday>) -> Self

Sets the birthday claim.

pub fn zoneinfo(&self) -> Option<&EndUserTimezone>

Returns the zoneinfo claim.

pub fn set_zoneinfo(self, zoneinfo: Option<EndUserTimezone>) -> Self

Sets the zoneinfo claim.

pub fn locale(&self) -> Option<&LanguageTag>

Returns the locale claim.

pub fn set_locale(self, locale: Option<LanguageTag>) -> Self

Sets the locale claim.

pub fn phone_number(&self) -> Option<&EndUserPhoneNumber>

Returns the phone_number claim.

pub fn set_phone_number(self, phone_number: Option<EndUserPhoneNumber>) -> Self

Sets the phone_number claim.

pub fn phone_number_verified(&self) -> Option<bool>

Returns the phone_number_verified claim.

pub fn set_phone_number_verified(
self,
phone_number_verified: Option<bool>
) -> Self

Sets the phone_number_verified claim.

pub fn address(&self) -> Option<&AddressClaim>

Returns the address claim.

pub fn set_address(self, address: Option<AddressClaim>) -> Self

Sets the address claim.

pub fn updated_at(&self) -> Option<DateTime<Utc>>

Returns the updated_at claim.

pub fn set_updated_at(self, updated_at: Option<DateTime<Utc>>) -> Self

Sets the updated_at claim.

pub fn standard_claims(&self) -> &StandardClaims<GC>

Returns the standard claims as a StandardClaims object.

pub fn additional_claims(&self) -> &AC

Returns additional user info claims.

pub fn additional_claims_mut(&mut self) -> &mut AC

Returns mutable additional user info claims.