Trait openidconnect::JsonWebKeyUse

pub trait JsonWebKeyUse: Debug + DeserializeOwned + Serialize + 'static {
    fn allows_signature(&self) -> bool;

    fn allows_encryption(&self) -> bool;
}

Allowed key usage.

Required methods

fn allows_signature(&self) -> bool

Returns true if the associated key may be used for digital signatures, or false otherwise.

fn allows_encryption(&self) -> bool

Returns true if the associated key may be used for encryption, or false otherwise.