Trait openidconnect::PrivateSigningKey

pub trait PrivateSigningKey<JS, JT, JU, K> where
    JS: JwsSigningAlgorithm<JT>,
    JT: JsonWebKeyType,
    JU: JsonWebKeyUse,
    K: JsonWebKey<JS, JT, JU>, {
    fn sign(
        &self, 
        signature_alg: &JS, 
        message: &[u8]
    ) -> Result<Vec<u8>, SigningError>;

    fn as_verification_key(&self) -> K;
}

Private or symmetric key for signing.

Required methods

fn sign(
&self,
signature_alg: &JS,
message: &[u8]
) -> Result<Vec<u8>, SigningError>

Signs the given message using the given signature algorithm.

fn as_verification_key(&self) -> K

Converts this key to a JSON Web Key that can be used for verifying signatures.